// src/utils/redirection.enum.ts
export enum RedirectionEnum {
  SEP = '/',
  HOME = '',
  STATUS = 'status',
  REGISTRATION_YOUNG = 'registration/13_18',
  REGISTRATION_ADULT = 'registration/19_26',
  REGISTRATION_STAFF = 'registration/volunteer',
  STATUE_YOUNG = 'statue/13_18',
  STATUE_ADULT = 'statue/19_26',
  LOGIN = 'login',
  ADMIN = 'admin-panel',
}